Always return float

+++ create-your-own-lang-with-rust/secondlang/src/lib.rs
@@ -45,12 +45,12 @@ pub use types::Type;
 pub use visitor::{AlgebraicSimplifier, ConstantFolder, ExprVisitor, PrettyPrinter};
 
 /// Convenience function to compile and run source code
-pub fn run(source: &str) -> Result<i64, String> {
+pub fn run(source: &str) -> Result<f64, String> {
     run_with_opts(source, false)
 }
 
 /// Compile and run with optional optimizations
-pub fn run_with_opts(source: &str, optimize: bool) -> Result<i64, String> {
+pub fn run_with_opts(source: &str, optimize: bool) -> Result<f64, String> {
     let mut program = parse(source)?;
     typecheck(&mut program)?;
